remove from_payload, more tests
